Added mass tasks adding
Added new activity to allow for mass adding of tasks through a string separated by new lines for each task.

Also added function to TaskAdapter to transform the newline string to the tasks string to keep the formatting in one place.

Fixed a few bugs related to entering and exiting activities with a no tasks

Added option to delete all tasks